requirements.txt
langchain
anthropic
pymongo
streamlit
pypdf
python-dotenv